# Reimplement JsonExpressionParser in terms of jsonparser 