--- a/pkg/logql/log/parser_test.go
+++ b/pkg/logql/log/parser_test.go
@@ -2,13 +2,12 @@ package log
 
 import (
 	"fmt"
+	"github.com/grafana/loki/pkg/logqlmodel"
 	"sort"
 	"testing"
 
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/require"
-
-	"github.com/grafana/loki/pkg/logqlmodel"
 )
 
 func Test_jsonParser_Parse(t *testing.T) {
@@ -1046,6 +1045,47 @@ func Test_PatternParser(t *testing.T) {
 			_, _ = pp.Process(0, tt.line, b)
 			sort.Sort(tt.want)
 			require.Equal(t, tt.want, b.LabelsResult().Labels())
+		})
+	}
+}
+
+func BenchmarkJsonExpressionParser(b *testing.B) {
+	simpleJsn := []byte(`{
+      "data": "Click Here",
+      "size": 36,
+      "style": "bold",
+      "name": "text1",
+      "hOffset": 250,
+      "vOffset": 100,
+      "alignment": "center",
+      "onMouseUp": "sun1.opacity = (sun1.opacity / 100) * 90;"
+    }`)
+	lbs := NewBaseLabelsBuilder().ForLabels(labels.Labels{}, 0)
+
+	benchmarks := []struct {
+		name string
+		p    Stage
+		line []byte
+	}{
+		{"json-expression", mustStage(NewJSONExpressionParser([]LabelExtractionExpr{
+			NewLabelExtractionExpr("data", "data"),
+			NewLabelExtractionExpr("size", "size"),
+			NewLabelExtractionExpr("style", "style"),
+			NewLabelExtractionExpr("name", "name"),
+			NewLabelExtractionExpr("hOffset", "hOffset"),
+			NewLabelExtractionExpr("vOffset", "vOffset"),
+			NewLabelExtractionExpr("alignment", "alignment"),
+			NewLabelExtractionExpr("onMouseUp", "onMouseUp"),
+		})), simpleJsn},
+	}
+	for _, bb := range benchmarks {
+		b.Run(bb.name, func(b *testing.B) {
+			b.ResetTimer()
+
+			for n := 0; n < b.N; n++ {
+				lbs.Reset()
+				_, result = bb.p.Process(0, bb.line, lbs)
+			}
 		})
 	}
 }