@MasslessParticle 2023-03-08
@dannykopping Here's @cyriltovena's original PR for JSON Parser: https://github.com/grafana/loki/pull/7723.

the `jsonparser` library helped somewhat. [this]() PR ought to buy us quite a bit -- at lease for metrics queries: https://github.com/grafana/loki/pull/8724