@MasslessParticle 2023-03-08
@dannykopping It's interesting to note that when the JSON payload is small and doesn't contain nested keys, the JSON Parser performs about as well as `logfmt`. [These](https://github.com/grafana/loki/pull/8724) benchmarks use a simpler JSON that @cyriltovena's